requirements.txt
numpy
pandas
scikit-learn
seqeval
gensim
python-crfsuite
pytorch-crf